Fix compilation error (#125)
* Remove plots from test

* Add continuous dynamics autodiff method to SpacecraftROE

This commit introduces a new method, `getContinuousDynamicsAutodiff`, to the `SpacecraftROE` class for computing continuous-time dynamics in ROE coordinates using automatic differentiation. Additionally, the state and control Jacobians have been updated to utilize this new method, enhancing the accuracy of the dynamics calculations. The changes improve the overall flexibility and precision of the spacecraft dynamics model within the CDDP framework.

* Fix bug in dynamics